feat(TweetManager): add hide-unhide reply methods
